Re-gen project with SDK 4.0.0 and 6.1

+++ maven-sdk/maven-sdk-tutorial/run.sh
@@ -24,33 +24,31 @@ start_acs() {
 }
 
 down() {
-    docker-compose -f $COMPOSE_FILE_PATH down
+    if [ -f $COMPOSE_FILE_PATH ]; then
+        docker-compose -f $COMPOSE_FILE_PATH down
+    fi
 }
 
 purge() {
-    docker volume rm maven-sdk-tutorial-acs-volume
-    docker volume rm maven-sdk-tutorial-db-volume
-    docker volume rm maven-sdk-tutorial-ass-volume
+    docker volume rm -f maven-sdk-tutorial-acs-volume
+    docker volume rm -f maven-sdk-tutorial-db-volume
+    docker volume rm -f maven-sdk-tutorial-ass-volume
 }
 
 build() {
-    docker rmi alfresco-content-services-maven-sdk-tutorial:development
-    docker rmi alfresco-share-maven-sdk-tutorial:development
-    $MVN_EXEC clean install -DskipTests=true
+    $MVN_EXEC clean package
 }
 
 build_share() {
     docker-compose -f $COMPOSE_FILE_PATH kill maven-sdk-tutorial-share
     yes | docker-compose -f $COMPOSE_FILE_PATH rm -f maven-sdk-tutorial-share
-    docker rmi alfresco-share-maven-sdk-tutorial:development
-    $MVN_EXEC clean install -DskipTests=true -pl maven-sdk-tutorial-share-jar
+    $MVN_EXEC clean package -pl maven-sdk-tutorial-share,maven-sdk-tutorial-share-docker
 }
 
 build_acs() {
     docker-compose -f $COMPOSE_FILE_PATH kill maven-sdk-tutorial-acs
     yes | docker-compose -f $COMPOSE_FILE_PATH rm -f maven-sdk-tutorial-acs
-    docker rmi alfresco-content-services-maven-sdk-tutorial:development
-    $MVN_EXEC clean install -DskipTests=true -pl maven-sdk-tutorial-platform-jar
+    $MVN_EXEC clean package -pl maven-sdk-tutorial-integration-tests,maven-sdk-tutorial-platform,maven-sdk-tutorial-platform-docker
 }
 
 tail() {
@@ -61,14 +59,25 @@ tail_all() {
     docker-compose -f $COMPOSE_FILE_PATH logs --tail="all"
 }
 
+prepare_test() {
+    $MVN_EXEC verify -DskipTests=true -pl maven-sdk-tutorial-platform,maven-sdk-tutorial-integration-tests,maven-sdk-tutorial-platform-docker
+}
+
 test() {
-    $MVN_EXEC verify -pl integration-tests
+    $MVN_EXEC verify -pl maven-sdk-tutorial-platform,maven-sdk-tutorial-integration-tests
 }
 
 case "$1" in
   build_start)
     down
     build
+    start
+    tail
+    ;;
+  build_start_it_supported)
+    down
+    build
+    prepare_test
     start
     tail
     ;;
@@ -99,6 +108,7 @@ case "$1" in
   build_test)
     down
     build
+    prepare_test
     start
     test
     tail_all
@@ -108,5 +118,5 @@ case "$1" in
     test
     ;;
   *)
-    echo "Usage: $0 {build_start|start|stop|purge|tail|reload_share|reload_acs|build_test|test}"
+    echo "Usage: $0 {build_start|build_start_it_supported|start|stop|purge|tail|reload_share|reload_acs|build_test|test}"
 esac
